Masked Span Task Game: "ChimpGPT"
The Masked Span Task Game used to measure Chimpanzee's memory in the Kyoto University research study in 2007 is back.
After seeing GPT3 and 4s creations of classics like Pong, the Chimpanzee study game inspired this recreation. This entire game was created with ChatGPT Plus by a non-coder. It is my  first app. While theres no GPT'ness about this game, I'm calling it ChimpGPT as a hat tip to both the original study and the amazing tool OpenAI has created.